backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
